PIM-4395: move cleaner in connector folder
