+++ src/Pim/Bundle/EnrichBundle/Connector/Item/MassEdit/MassEditTemporaryFileCleaner.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\MassEditAction\Cleaner;
+namespace Pim\Bundle\EnrichBundle\Connector\Item\MassEdit;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
