+++ src/Pim/Bundle/EnrichBundle/Connector/Step/MassEditStep.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\EnrichBundle\Connector\Step;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Step\AbstractStep;
 use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
-use Pim\Bundle\EnrichBundle\MassEditAction\Cleaner\MassEditTemporaryFileCleaner;
+use Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\MassEditTemporaryFileCleaner;
 use Pim\Bundle\EnrichBundle\Step\MassEditRemoveTemporaryMediaStep;
 
 /**
